requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
config==0.5.1
decouple==0.0.7
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
lxml==4.9.2
MarkupSafe==2.1.3
multidict==6.0.4
nltk==3.8.1
numpy==1.24.3
Pillow==9.5.0
pysummarization==1.1.9
python-decouple==3.8
python-dotenv==1.0.0
python-pptx==0.6.21
regex==2023.6.3
requests==2.31.0
SQLAlchemy==2.0.16
tqdm==4.65.0
typing_extensions==4.6.3
urllib3==2.0.3
Werkzeug==2.3.6
wtf==0.1
WTForms==3.0.1
XlsxWriter==3.1.2
yarl==1.9.2
